Fix keychain accessibility issue
I began to notice instances where the widget would show an error message on the lock screen even though the user was logged in. I narrowed down the issue to keychain accessibility. The Locksmith framework didn't allow me to set `kSecAttrAccessibleAfterFirstUnlock` which allows the keychain to be accessed from the background (assuming the phone has been unlocked before). I have thus migrated to the KeychainAccess framework